avoid crash when trying to cleanup message in deleted channel - update Best-in-Class Hammer flavor text

+++ source/buttons/readyitem.js
@@ -84,7 +84,9 @@ module.exports = new ButtonWrapper(mainId, 3000,
 				console.error(error)
 			}
 		}).finally(() => {
-			interaction.deleteReply();
+			if (interaction.channel) { // prevent crash if channel is deleted before cleanup
+				interaction.deleteReply();
+			}
 		});
 	}
 );
