+++ source/commands/set-favorite/archetype.js
@@ -35,7 +35,9 @@ async function executeSubcommand(interaction, ...[player]) {
 			console.error(error);
 		}
 	}).finally(() => {
-		interaction.deleteReply();
+		if (interaction.channel) { // prevent crash if channel is deleted before cleanup
+			interaction.deleteReply();
+		}
 	});
 };
