+++ source/commands/set-favorite/pet.js
@@ -40,7 +40,9 @@ async function executeSubcommand(interaction, ...[player]) {
 			console.error(error);
 		}
 	}).finally(() => {
-		interaction.deleteReply();
+		if (interaction.channel) { // prevent crash if channel is deleted before cleanup
+			interaction.deleteReply();
+		}
 	});
 };
