+++ source/selects/treasure.js
@@ -129,7 +129,9 @@ module.exports = new SelectWrapper(mainId, 2000,
 						console.error(error);
 					}
 				}).finally(() => {
-					interaction.deleteReply();
+					if (interaction.channel) { // prevent crash if channel is deleted before cleanup
+						interaction.deleteReply();
+					}
 				});
 				break;
 		}
